pyproject.toml
[project]
name = "emotion-detection-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
